First stage change to switch to single process shared
Original behavior was spawining a new process each time a client
connects to the server, and sending to this client it's own process
context and contents.

This changes this behavior, now when server starts, the process
requested is started at the same time. When a client connect it attaches
itself to this process and can interract with it. Any extra client are
attached too, and share everything.

This makes this working like a console multiplexing over webui